cleaned up github repository: nested Chart.java and DatabaseInterface.java in a src folder and deleted "Project Analysis Initial brainstorm Todd.docx"
